backend/requirements.txt
fastapi==0.111.0
pydantic==2.7.1
SQLAlchemy==2.0.30
uvicorn==0.29.0
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1